third_party/blink/web_tests/http/tests/security/subresourceIntegrity/subresource-integrity-style-no-cors-no-xorigin-empty-integrity.html

<!DOCTYPE html>
<head>
    <title>SRI on cross origin CORS disabled style, without CORS fetch</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script>
    setup({single_test: true});

    function styleLoad() {
        done();
    }
    function styleError() {
        assert_unreached("Style did not load.");
        done();
    }
    </script>
    <style>.id1 { background-color: red }</style>
    <link rel="stylesheet"
          onload="styleLoad()"
          onerror="styleError();"
          integrity=""
          href="http://localhost:8000/security/resources/cors-style.php?cors=false"></link>
</head>
<body>
    <script>
        window.onload = function () {
            var ele = document.getElementById('id1');
            var divBackground = window.getComputedStyle(ele, null).getPropertyValue('background-color');
            assert_equals(divBackground, "rgb(255, 0, 0)");
            done();
        };
    </script>
    <div id="id1" class="id1"></div>
</body>
</html>
